move tests shared functions and classes to /tests/utils (#338)
* move tests shared functions to /tests/utils

* move FakeGTP to /utils/test

* move FakeSocket and FakeAPI shared classes to /test/utils

* move other tests shared functions to /tests/utils/

requireUncached will be used in other PRs, for example to test
config.js in #320

* rename to assignConfigArguments

* rename to resetToNew

* add newline at end of files

* remove config.DEBUG from tests

we don't use console logging for tests, only in the real gtp2ogs

if we want to know the value of a variable at any time, we can use
our own debugger (ex: vs code's debugger)

* move more shared functions to /test/utils

some of these functions will be used in config.test.js and other
future test files

* add newline at end of files
