+++ test/test.js
@@ -1,314 +1,27 @@
 // vim: tw=120 softtabstop=4 shiftwidth=4
 
 const assert = require('assert');
+const child_process = require('child_process');
+const fs = require('fs');
+const https = require('https');
 const sinon = require('sinon');
 
-const connection = require('../connection');
-const config = require('../config');
-const { console } = require('../console');
+const { base_active_game } = require('./utils/base_active_game');
+const { base_challenge } = require('./utils/base_challenge');
+const { base_gamedata } = require('./utils/base_gamedata');
+const { FakeAPI } = require('./utils/FakeAPI');
+const { FakeGTP } = require('./utils/FakeGTP');
+const { FakeSocket } = require('./utils/FakeSocket');
+const { getNewConfig } = require('./utils/getNewConfig');
+const { stub_console } = require('./utils/stub_console');
+
 const { Bot } = require('../bot');
+const config = getNewConfig();
+const connection = require('../connection');
+const { console } = require('../console');
 
-const child_process = require('child_process');
-const https = require('https');
-
-const fs = require('fs')
-
-const stream = new require('stream');
-
-config.DEBUG = true;
-config.apikey = 'deadbeef';
-config.host = 'test';
-config.port = 80;
-config.username = 'testbot';
-
-config.bot_command = ['gtp-program', '--argument'];
 config.timeout = 0; // needed for test.js
 config.corrqueue = false; // needed for test.js
-
-// Fake a socket.io-client
-class FakeSocket {
-    constructor() {
-        this.on_callbacks = {};
-        this.emit_callbacks = {};
-    }
-
-    on(ev, cb) {
-        console.log('client subscribe: ' + ev)
-        this.on_callbacks[ev] = cb;
-    }
-
-    inject(ev, data) {
-        console.log('client on(' + ev + ')')
-        this.on_callbacks[ev](data);
-    }
-
-    emit(ev, data, cb) {
-        if (config.DEBUG) {
-            console.log('client: ' + ev);
-        }
-        let ret;
-        if (this.emit_callbacks[ev]) {
-            ret = this.emit_callbacks[ev](data);
-        }
-        if (cb) {
-            cb(ret);
-        }
-    }
-
-    on_emit(ev, cb) {
-        console.log('server subscribe: ' + ev);
-        this.emit_callbacks[ev] = cb;
-    }
-}
-
-// Fake http/https request 
-class FakeAPI {
-    constructor() {
-        this.callbacks = {};
-        this.request = this.request.bind(this);
-    }
-
-    on_path(path, cb) {
-        this.callbacks[path] = cb;
-    }
-
-    request(options, cb) {
-        let response = '';
-        console.log('api ' + options.path);
-        if (this.callbacks[options.path]) {
-            response = this.callbacks[options.path](options);
-        }
-        cb({
-            statusCode: 200,
-            setEncoding: () => {},
-            on: (ev, cb) => {
-                if (ev === 'data') {
-                    cb(response);
-                }
-                if (ev === 'end') {
-                    cb();
-                }
-            },
-        });
-        return {
-            on: () => {},
-            write: () => {},
-            end: () => {},
-        };
-    }
-}
-
-// Fake GTP child_process (spwan)
-class FakeGTP {
-    constructor() {
-        this.pid = 100;
-        this.callbacks = {};
-        this.cmd_callbacks = {};
-        this.stderr = new stream.Readable({ read: () => {}});
-        this.callbacks.stderr = (data) => this.stderr.emit('data', data);
-
-        this.stdout = { on: (_, cb) => {
-            this.callbacks.stdout = cb;
-        }};
-        this.stdin = {
-            on: (_, cb) => {
-                this.callbacks.stdin = cb;
-            },
-            end: () => {},
-            write: (data) => {
-                if (config.DEBUG) {
-                    console.log('STDIN: ', data.trim());
-                }
-                let cmd = data.trim().split(' ')[0];
-                if (this.cmd_callbacks[cmd]) {
-                    this.cmd_callbacks[cmd](data.trim());
-                } else {
-                    this.gtp_response('');
-                }
-            }
-        };
-    }
-
-    on(ev, cb) {
-        this.callbacks[ev] = cb;
-    }
-
-    on_cmd(cmd, cb) {
-        console.log('GTP: ', cmd);
-        this.cmd_callbacks[cmd] = cb;
-    }
-
-    gtp_response(data) {
-        this.callbacks.stdout('= ' + data + "\n\n");
-    }
-
-    gtp_error(data) {
-        this.callbacks.stdout('? ' + data + "\n\n");
-    }
-
-    exit(code) {
-        if (this.callbacks.exit) {
-            this.callbacks.exit({ code: code, signal: null });
-        }
-    }
-
-    kill() {
-        this.exit(1);
-    }
-}
-
-function base_challenge(overrides) {
-    let base = {
-        type: 'challenge',
-        challenge_id: 1,
-        player_id: 1,
-        game_id: 1,
-        id: '1', // GUID
-        challenger_color: 'white',
-        user: {
-            id: 2,
-            username: 'human',
-            professional: false,
-            ranking: 10.0,
-        },
-        rules: 'chinese',
-        ranked: true,
-        width: 19,
-        height: 19,
-        handicap: 0,
-        komi: 7.5,
-        time_control: {
-            speed: 'live',
-            system: 'fischer',
-            time_control: 'fischer',
-            initial_time: 120,
-            max_time: 300,
-            time_increment: 30,
-            pause_on_weekends: false,
-        },
-        disable_analysis: false,
-        aga_rated: false,
-        aux_delivered: 0,
-        read: 0,
-        timestamp: 0,
-        read_timestamp: 0,
-    }
-    return Object.assign({}, base, overrides);
-}
-
-function base_active_game(overrides) {
-    let base = {
-        id: 1,
-        phase: 'play',
-        name: 'Friendly Match',
-        player_to_move: 1,
-        time_per_move: 89280,
-        width: 19,
-        height: 19,
-        move_number: 0,
-        paused: 0,
-        private: false,
-        black: {
-            id: 1,
-            username: 'testbot',
-            rank: 10,
-            professional: false,
-            accepted: false,
-        },
-        white: {
-            id: 2,
-            username: 'human',
-            rank: 10,
-            professional: false,
-            accepted: false,
-        },
-    }
-    return Object.assign({}, base, overrides);
-}
-
-function base_gamedata(overrides) {
-    let base = {
-        game_id: 1,
-        game_name: 'Friendly Match',
-        phase: 'play',
-        komi: 7.5,
-        handicap: 0,
-        width: 19,
-        height: 19,
-        private: false,
-        ranked: false,
-        rules: 'chinese',
-        time_control: {
-            system: 'fischer',
-            pause_on_weekends: false,
-            time_control: 'fischer',
-            initial_time: 120,
-            max_time: 300,
-            time_increment: 30,
-            speed: 'live',
-        },
-        start_time: 0,
-        clock: {
-            game_id: 1,
-            current_player: 1,
-            black_player_id: 1,
-            white_player_id: 2,
-            title: 'Friendly Match',
-            last_move: 0,
-            expiration: 0, // TODO
-            black_time: { thinking_time: 30, skip_bonus: false },
-            white_time: { thinking_time: 30, skip_bonus: false },
-            start_mode: true,
-        },
-        initial_player: 'black',
-        black_player_id: 1,
-        white_player_id: 2,
-        players: {
-            black: {
-                id: 1,
-                username: 'testbot',
-                rank: 10,
-                professional: false,
-            },
-            white: {
-                id: 2,
-                username: 'human',
-                rank: 10,
-                professional: false,
-            },
-        },
-        moves: [],
-        meta_groups: [],
-        history: [],
-        initial_state: { black: '', white: '' },
-        pause_on_weekends: false,
-        disable_analysis: false,
-        allow_self_capture: false,
-        automatic_stone_removal: false,
-        free_handicap_placement: true,
-        aga_handicap_scoring: false,
-        allow_ko: false,
-        allow_superko: false,
-        superko_algorithm: 'ssk',
-        score_territory: true,
-        score_territory_in_seki: true,
-        score_stones: true,
-        score_handicap: true,
-        score_prisoners: false,
-        score_passes: true,
-        white_must_pass_last: false,
-        opponent_plays_first_after_resume: false,
-        strict_seki_mode: false,
-        original_disable_analysis: false,
-    };
-    return Object.assign({}, base, overrides);
-}
-
-function stub_console() {
-    sinon.stub(console, 'log');
-    sinon.stub(console, 'debug');
-    sinon.stub(console, 'error');
-}
 
 afterEach(function () {
     sinon.restore();
@@ -817,5 +530,3 @@ describe("Pv should work", () => {
         conn.terminate();
     }
 });
-
-module.exports = { FakeSocket, FakeAPI, base_challenge };
