+++ test/utils/getNewConfig.js
@@ -1,0 +1,9 @@
+const { assignConfigArguments } = require('./assignConfigArguments.js');
+
+function getNewConfig() {
+    const config = require('../../config');
+    assignConfigArguments(config);
+    return config;
+}
+
+exports.getNewConfig = getNewConfig;
