+++ test/utils/getNewConnectionUncached.js
@@ -1,0 +1,8 @@
+const { requireUncached } = require('./requireUncached');
+
+function getNewConnectionUncached() {
+    const connection = requireUncached('../../connection');
+    return connection;
+}
+
+exports.getNewConnectionUncached = getNewConnectionUncached;
